Remove header_parser (API Change)
