+++ test/http/doc_http_samples.cpp
@@ -256,134 +256,6 @@ public:
     }
 
     //--------------------------------------------------------------------------
-#if 0
-    // VFALCO This is broken
-    /*
-        Efficiently relay a message from one stream to another
-    */
-    template<
-        bool isRequest,
-        class SyncWriteStream,
-        class DynamicBuffer,
-        class SyncReadStream>
-    void
-    relay(
-        SyncWriteStream& out,
-        DynamicBuffer& b,
-        SyncReadStream& in)
-    {
-        flat_buffer buffer{4096}; // 4K limit
-        header_parser<isRequest, fields> parser;
-        serializer<isRequest, buffer_body<
-            typename flat_buffer::const_buffers_type>,
-                fields> ws{parser.get()};
-        error_code ec;
-        do
-        {
-      	    auto const state0 = parser.state();
-            auto const bytes_used =
-                read_some(in, buffer, parser, ec);
-            BEAST_EXPECTS(! ec, ec.message());
-            switch(state0)
-            {
-            case parse_state::header:
-            {
-                BEAST_EXPECT(parser.is_header_done());
-                for(;;)
-                {
-                    ws.write_some(out, ec);
-                    if(ec == http::error::need_more)
-                    {
-                        ec = {};
-                        break;
-                    }
-                    if(ec)
-                        BOOST_THROW_EXCEPTION(system_error{ec});
-                }
-                break;
-            }
-
-            case parse_state::chunk_header:
-            {
-                // inspect parser.chunk_extension() here
-                if(parser.is_done())
-                    boost::asio::write(out,
-                        chunk_encode_final());
-                break;
-            }
-
-            case parse_state::body:
-            case parse_state::body_to_eof:
-            case parse_state::chunk_body:
-            {
-                if(! parser.is_done())
-                {
-                    auto const body = parser.body();
-                    boost::asio::write(out, chunk_encode(
-                        false, boost::asio::buffer(
-                            body.data(), body.size())));
-                }
-                break;
-            }
-          
-            case parse_state::complete:
-                break;
-            }
-            buffer.consume(bytes_used);
-        }
-        while(! parser.is_done());
-    }
-
-    void
-    testRelay()
-    {
-        // Content-Length
-        {
-            test::string_istream is{ios_,
-                "GET / HTTP/1.1\r\n"
-                "Content-Length: 5\r\n"
-                "\r\n" // 37 byte header
-                "*****",
-                3 // max_read
-            };
-            test::string_ostream os{ios_};
-            flat_buffer b{16};
-            relay<true>(os, b, is);
-        }
-
-        // end of file
-        {
-            test::string_istream is{ios_,
-                "HTTP/1.1 200 OK\r\n"
-                "\r\n" // 19 byte header
-                "*****",
-                3 // max_read
-            };
-            test::string_ostream os{ios_};
-            flat_buffer b{16};
-            relay<false>(os, b, is);
-        }
-
-        // chunked
-        {
-            test::string_istream is{ios_,
-                "GET / HTTP/1.1\r\n"
-                "Transfer-Encoding: chunked\r\n"
-                "\r\n"
-                "5;x;y=1;z=\"-\"\r\n*****\r\n"
-                "3\r\n---\r\n"
-                "1\r\n+\r\n"
-                "0\r\n\r\n",
-                2 // max_read
-            };
-            test::string_ostream os{ios_};
-            flat_buffer b{16};
-            relay<true>(os, b, is);
-        }
-    }
-#endif
-    
-    //--------------------------------------------------------------------------
 
     void
     run()
@@ -395,7 +267,6 @@ public:
         doWriteStdStream();
         doCustomParser();
         doHEAD();
-
         doDeferredBody();
     }
 };
